+++ docs/root/intro/version_history.rst
@@ -8,6 +8,7 @@ Version history
 * buffer: fix vulnerabilities when allocation fails
 * config: removed deprecated_v1 sds_config from :ref:`Bootstrap config <config_overview_v2_bootstrap>`.
 * redis: added :ref:`success and error stats <config_network_filters_redis_proxy_per_command_stats>` for commands.
+* redis: added :ref:`latency stats <config_network_filters_redis_proxy_per_command_stats>` for commands.
 * tls: enabled TLS 1.3 on the server-side (non-FIPS builds).
 
 1.9.0